pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "weightborne"
version = "0.1.0"
description = "A Flask web app for tracking weight with analytics and forecasting."
requires-python = ">=3.12"

dependencies = [
    "Flask==3.1.3",
    "Flask-Login==0.6.3",
    "Flask-Mail>=0.10",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Migrate",
    "Flask-Limiter>=3.5",
    "Flask-WTF>=1.2",
    "SQLAlchemy==2.0.48",
    "Werkzeug==3.1.7",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["migrations/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
